Add scripts for counting merges, conflicts
all-merges gives a list of all the merges, one line
per merge.  Format is
<id1> <id2> ... ==> <merged id>

conflicting-merges prints "conflict" for each conflict
found.  This needs to be fixed up to print conflict ids
and not other stuff, and to reuse data computed by
all-merges.

get-merge-data calls the other two scripts and prints
the number of merges and number of conflicts.