Miscellaneous fixes to auto-merge
1. Remove unused +limit-program-path+ constant.

2. Pass :max-time/:max-evals properly to the `resolve` method.

3. Fix `resolve` method signature to match callers.

4. Unquote `test` variable properly in `resolve` method.

5. Set the *target-function-p* variable properly in the `resolve`
method.  Ensure this a function and not nil.

6. The `resolve` method returns the 'best' variant.

7. Output single-file programs without error.

8. Allow setting --out-dir via command line.

9. Update argument-multiplier for args without a short name
(e.g. `split-lines`).

10. Add ignorable declarations for `split-lines`.

11. Simplify the `converge` method implementation.  Remove
unneccessary let bindings.

12. Semantically neutral changes to the `resolve-conflict` method.
I find the method more readable in this form.

13. Remove redundant conflict resolution strategy.
The :CB conflict resolution strategy was a duplicate of
:C1 or :C2 depending on the result of the randomization shuffle.

14. Allow the client to use evolution to resolve the merge if they
pass `--evolve`.  The evolutionary loop currently does not perform
satisfactorily as (1) hard errors are returned and (2) mutations
and crossover are not limited to those areas of the genome
near the merge conflicts.

15. Allow test scripts have multiple test cases.

Add a --num-tests parameter to allow for execution of the test
script multiple times.  Utilize the same signature as Bug-Injector
where we pass the test script and a placeholder, ~d, for the test
number.  The `test.sh` script bundled with the GCD program will
only operate properly if a test number is passed as the second
paramater.

This change also updates the evolutionary loop to use a non-numeric
fitness.  Instead, the fitness is a list where each element
corresponds to running the test script on a given test number.
The non-numeric fitness is utilized for lexicase selection.