Change to `ast-patch-conflict-action` special case
We will return the conflict node further below in the tree if
such a node exists for the special case where one conflict
action is :SAME and the other is :RECURSE.  Otherwise, we will
create a conflict node as usual from the current node.

Previously, a case could occur where a conflict node would not
be created for conflicts where the underlying edit was a string
manipulation such as :insert-sequence or :delete-sequence and
not an AST change such as :insert or :delete.