Change how ast-diff works on raw sexprs
Previously these were handled as-is in the machinery,
but now they're transformed to SIMPLE-LISP-AST structures
and are treated like any other ast (except that print-diff
does not work on them, as it didn't work on raw lists before.)

This changes greatly simplifies the internals.

Make AST-PATCH have an internal version AST-PATCH* that is called
recursively, like what was done to AST-DIFF.

Add more tests for coverage.