Add support for handling inserted/deleted files
In this case, we add conflict ASTs for the inserted
deleted files at the AST root.

To support this behavior, an auto-mergeable mixin class
was added to allow for auto-merge specific patching
behavior.  For auto-mergeable projects, we create
these conflict ASTs for the inserted and deleted
files at the AST root.

Further, the auto-merge crossover implementation
and other overrides of existing software object
behavior for auto-merge were added as specializations
on auto-mergeable software objects.