Update for use of `file-w-attributes` mixin in JS
JavaScript software objects now utilize the `file-w-attributes` mixin
to preserve file permissions instead of a workaround which checked
for a hashbang at the beginning of the script to determine if the
file is executable.  To preserve the attributes slot, we copy the
software object when performing a three way merge and update the
`ast-root` field with the result of the converge algorithm.